Actualizacion ej2

+++ ejercicios-java/tema03/Ejercicio2.java
@@ -7,9 +7,9 @@ public class Ejercicio2 {
   
   System.out.println("--Conversor de euros a pesetas--");
   System.out.printf("Escriba a continuación los euros que quiere pasar a pesetas: ");
-  int euros = Integer.parseInt(System.console().readLine());
+  double euros = Double.parseDouble(System.console().readLine());
   
-  double eurosAPesetas = euros * 166.386;
+  int eurosAPesetas = (int)(euros * 166.386);
   
   System.out.println(" -----------------------------");
   System.out.println(" " + euros + " euros son " + eurosAPesetas + " pesetas");
